TASK FINALLY FINISHED, CHESSBOTS BEGINS NOW!!!!!
