+++ src/server/command/command.ts
@@ -1,4 +1,4 @@
-import { gamePaused } from "../api/managers";
+import { gamePaused } from "../api/pauseHandler";
 import { robotManager } from "../robot/robot-manager";
 
 /**
